NETOBSERV-505 NOO fails to build with go1.18 and above
Use 'go install' to install the executable.  Note that it still calls
'go get' to update dependencies in go.mod.  This stopped working in
go1.18 because of https://go.dev/doc/go-get-install-deprecation.

In making this change, kustomize broke and no longer installed.  The problem
is documented in https://github.com/kubernetes-sigs/kustomize/issues/3618.
The solution is to use at least v4.5.2.  kustomize was upgraded to 4.5.7
from 3.8.8 and everything still worked.