Fix remaining Makefile issues + annoying bundle date
After the recent reworking on the Makefile, there were still some
pending issues. E.g. we introduced a "deploy-latest" target but it did
not allow to override the operator image.

This is rollbacking part of the last changes:
- deploy-latest is removed, we can run "make deploy" as we used to do
  before
- make deploy now uses again "main" images by default, as before
- keeping the behaviour that Mohamed introduced, "make bundle" keeps
  using the provided IMG
- UNLIKE before: running "make bundle" alone will NOT update the bundle
  so it is ready for commit; For that, a new target "make update-bundle"
is introduced
- "make bundle" does not generate the creation date as it used to do.
  This is now done via "make bundle-build". It means that RELEASE
SCRIPTS HAVE TO BE UPDATED so they generate the creation date.