NETOBSERV-909 fix servicemonitor & prom rule reconcile (#290)
* NETOBSERV-909 fix servicemonitor & prom rule reconcile

We need to be careful about not tying lifecycle of different objects
together, as this would be assuming a certain state, which creates "holes"
in the reconcile logic - we must avoid as much as possible any
assumption about the current state

* codecov no fail CI

Codecov fails too often to upload. Make it non mandatory
Also increase the timeout for some tests

* Remove all "static" initializations

They are also flawed: we should not assume they are created once for all
when flowcollector is created.
E.g. of how it could go wrong:
1. an old version of an operator is installed, with a flowcollector
2. an upgrade is done, which introduces a new "static" resource
=> that new resource would not be created, since the FlowCollector
object was already present hence did not trigger these "static init"
things