NETOBSERV-963, revert most of cert watching
Reverting most of certificate watching (NETOBSERV-684) as it generates a
lot of pods restart.

We did not necessarily have to do this certificate watching as
CM/secrets are updated within volumes.

We might however monitor carefully if new (or old) problems arise,
potentially due to the kubelet sync delay for updating volumes

And also make sure certificates aren't cached in our different workloads