NETOBSERV-1076 Fix CR stuck in Updating state
The problem was that copying certificates across namespaces was done on
every reconciliation call, with or without digest change. This is not
a critical issue as it would only trigger a no-op update, but it was
screwing up our "updating" state detection.

There's actually two changes in this commit:
1. Fixing the "Updating" flag, so that it's not set when an update is
   actually performed (we need to fetch back the object (from cache) and
check if the resource version was actually modified)
2. Fixing the certificate watching process to avoid unnecessary tries to
   update

Also:
- Add / fix tests
- Modify some logs to make important information more visible
- Makefile change for building manifest without error (unrelated)