Avoid enabling namespaces and workloads metrics tags
Namespace-based metrics are included in workload's ones, so this would
result in unnecesary metrics creation when both of them are enabled

So namespaces tags is now disabled by default; to still show
namespace-based metrics in dashboard, we need to use the workload
metrics aggregated by namespace. So We need to duplicate the related
rows in the dashboard.