NETOBSERV-1075 Dashboard feedback: split infra / applications (#388)
* flow metrics dashboard

* svc monitor check

* rebase fix

* tied metrics / dashboards

* infra_health_dashboard flows overhead

Co-authored-by: Joel Takvorian <joel.takvorian@qaraywa.net>

* fine grained tags

* Update bundle

* Fix linter issues

* Avoid enabling namespaces and workloads metrics tags

Namespace-based metrics are included in workload's ones, so this would
result in unnecesary metrics creation when both of them are enabled

So namespaces tags is now disabled by default; to still show
namespace-based metrics in dashboard, we need to use the workload
metrics aggregated by namespace. So We need to duplicate the related
rows in the dashboard.

* Update metrics definitions

* Inject namespace for infra filters

Changed for a programmatic approach, to handle more easily json
manipulation and avoid the very large (and very redundant) dashboard json

Also change from using topk5 to topk10 as it renders better charts (less
chunks in timeseries)

---------

Co-authored-by: Joel Takvorian <joel.takvorian@qaraywa.net>
Co-authored-by: Joel Takvorian <jtakvori@redhat.com>