NETOBSERV-1955: fix wrong metrics for conversation tracking (#1096)
* NETOBSERV-1955: fix wrong metrics for conversation tracking

This is fixing wrong metrics that are due to accumulating bytes/packets
counters from heartbits, which are not relevant for accumulation.

However, conversation tracking is still bad at providing accurate
metrics for long-standing connections, as the ConnectionEnded event may
be very long to come. So this is going to remain as a known limitation
of Conversation Tracking.

* update helm