Deploy FLP as a service
This is an intermediate alternative between the Kafka mode and the
Direct mode, that is more suitable for quick install on large clusters
(Kafka mode is a more complex setup, whereas Direct mode isn't suitable
on large clusters due to the memory consumption of FLP)

To use it, set `deploymentModel` to `Service`.

There are potential caveat to check:
- Without sticky session, no guarantee that the agents talk to the same
  FLP instance. I don't think it's an issue in the nominal case, but
might be a problem for conversation tracking?