[backport 1.10] NETOBSERV-2439: New "part-of" label + limit netpol to netobserv pods (#2088)
* New "part-of" label on all pods / resources (#2078)

Introduce the "part-of" label to designate by labels all resources part
of the netobserv operator.

This aims to simplify user workflow if they need to do something related
to all netobserv resources; such as configuring a network policy that
only target netobserv pods, for instance.

* NETOBSERV-2439: limit netpol to netobserv pods

Also revert the validation webhook on netpol, which becomes useless